I added a ton of stuff.
1) lots of comments

2) table_maker which contains the code Stephano and David were using to generate tables.
We will need this in August and I'm afraid it is going to get lost.

3) graph_example functions have been  moved to graph_builder.

There are a lot of them including a shortest length
isogeny function. This needs to be tests. I'm actually not sure if graph_example can just be loaded and ran as-is.
It should be ok with minor modifications.

The output from the 02/02/2019 LMFDB meeting is on a sage worksheet that both Edgar and I are on.

4) graph_example still needs the horizontal isogenies and backwards isogenies.